`@remotion/serverless`: Extract entire Lambda runtime into serverless runtime
