+++ packages/lambda/src/shared/constants.ts
@@ -15,7 +15,6 @@ export const MAX_TIMEOUT = 900;
 export const DEFAULT_FRAMES_PER_LAMBDA = 20;
 
 export const BINARY_NAME = 'remotion lambda';
-export const COMMAND_NOT_FOUND = 'Command not found';
 export const DEFAULT_REGION: AwsRegion = 'us-east-1';
 export const DEFAULT_MAX_RETRIES = 1;
