+++ packages/serverless/src/is-warm.ts
