+++ packages/serverless/src/print-logging-grep-helper.ts
@@ -1,6 +1,6 @@
 import type {LogLevel} from '@remotion/renderer';
 import {RenderInternals} from '@remotion/renderer';
-import type {ServerlessRoutines} from '@remotion/serverless/client';
+import type {ServerlessRoutines} from './constants';
 
 export type PrintLoggingHelper = (
 	type: ServerlessRoutines,
